fix some division errors

+++ pydl/pydlutils/misc/tests/test_struct_print.py
@@ -5,8 +5,8 @@ def test_struct_print():
     from .. import struct_print
     n = 20
     slist = np.zeros(n,dtype=[('PLATE','i4'),('MJD','i4'),('FIBERID','i4'),('RA','f8'),('DEC','f8'),('MATCHRAD','f4'),('RERUN','S8')])
-    slist['RERUN'][0:n/2] = 'v1_2_3'
-    slist['RERUN'][n/2:n] = 'v1_3_34'
+    slist['RERUN'][0:int(n/2)] = 'v1_2_3'
+    slist['RERUN'][int(n/2):n] = 'v1_3_34'
     slist['PLATE'] = np.random.random_integers(10,10000,(n,))
     slist['MJD'] = np.random.random_integers(51000,56000,(n,))
     slist['FIBERID'] = np.random.random_integers(1,1000,(n,))
